discogs_credits: extract dispatch orchestrator
Issue #32 — `instantFillRelationships` (~670 lines) moves to
`src/dispatch.js`. The main "walk the resolved Discogs JSON and feed
every credit into MB's relationship editor" loop, with track/work
routing, dedup against MB's existing rels (issue #14), and the edit-note
update on completion.

Imports the full stack of already-extracted modules:
  log, constants, storage, api-discogs, api-mb, editor-state,
  edit-note, data/entity-map, data/work-only-rels.

~670 lines moved. Smoke (Spiritual Jazz: 95 staged rels) green.
